third_party/WebKit/LayoutTests/fast/css/pseudo-default-001.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
<body>
<script src="../../resources/js-test.js"></script>
<form method="get">
<input name="victim" type="submit" value="Submit" id="enabledSubmit"/>
</form>
<form method="get">
<input name="victim" type="submit" value="Submit" disabled id="disabledSubmit"/>
</form>
<script>
description("This test performs a simple check for the :default CSS selector.");

var v = document.querySelectorAll(":default");
shouldBe("v.length", "2");
shouldBeEqualToString("v[0].id", "enabledSubmit");
shouldBeEqualToString("v[1].id", "disabledSubmit");
</script>
</body>
</html>
