third_party/blink/web_tests/fast/css/pseudo-default-001-expected.txt

This test performs a simple check for the :default CSS selector.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS v.length is 2
PASS v[0].id is "enabledSubmit"
PASS v[1].id is "disabledSubmit"
PASS successfullyParsed is true

TEST COMPLETE
